refactor: ZAttr classes
