+++ pom.xml
@@ -502,7 +502,7 @@
       <dependency>
         <groupId>com.zextras</groupId>
         <artifactId>mailbox-attribute-manager</artifactId>
-        <version>4.30.0</version>
+        <version>4.31.0-SNAPSHOT</version>
       </dependency>
       <dependency>
         <groupId>com.zextras</groupId>
