+++ store/src/main/java/com/zimbra/cs/account/MailTargetEntry.java
@@ -1,0 +1,15 @@
+package com.zimbra.cs.account;
+
+import java.util.Map;
+
+public abstract class MailTargetEntry extends MailTarget {
+
+	public MailTargetEntry(String name, String id, Map<String, Object> attrs,
+			Map<String, Object> defaults, Provisioning prov) {
+		super(name, id, attrs, defaults, prov);
+	}
+
+	public Provisioning getProvisioningInternal() {
+		return null;
+	}
+}
