+++ store/src/main/java/com/zimbra/cs/account/MailTargetEntryAttrs.java
@@ -1,0 +1,5 @@
+package com.zimbra.cs.account;
+
+public class MailTargetEntryAttrs {
+
+}
